django_file/requirements.txt
#python3.10.14
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
Brotli==1.0.9
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.3
cryptography==42.0.5
Deprecated==1.2.14
Django==4.1
django-celery-beat==2.6.0
django-redis==5.4.0
django-redshift-backend==4.0.0
django-timezone-field==7.0
djangorestframework==3.14.0
dnspython==2.6.1
email_validator==2.1.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
Jinja2==3.1.4
kombu==5.3.7
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==4.3.4
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
pandas
geopy